adicionado paginacion

+++ application/usecases/usecase_products.py
@@ -1,17 +1,56 @@
 from fastapi import HTTPException, status
+from fastapi.responses import JSONResponse
 from domain.interfaces.repositories.repository_products import ProductRepository as repository
+from domain.interfaces.business_logic.business_logic_products import ProductsBusinessLogic as business_logic
 from domain.interfaces.repositories.repository_products import Session, get_session, Depends, List
 from domain.interfaces.repositories.repository_products import get_schema, post_schema, put_schema, patch_schema
 from domain.interfaces.repositories.repository_products import model
+from configuration.end_points.products import BASE_PATH
 from configuration.log.logging import log_api
 
-class ProductsUseCases(repository):
-     
+class ProductsUseCases(repository, business_logic):
+    
+    def get_response_format(self, data, offset: int = 0, limit: int = 100, page_number : int = 1, page_size : int = 10):
+        
+        if type(data) == list:
+            import math
+            start = (page_number - 1) * page_size
+            end = start + page_size
+            total_data = len(data)
+            total_pages = math.ceil(total_data / page_size)
+            
+            response_data = {
+                
+                'data':data[start:end],            
+                'count':page_size,
+                'total':len(data),
+                'page': f'{page_number} from {total_pages}',
+                'pagination':{
+                            'previous':None,
+                            'next':None
+                            }
+            }
+                        
+            if page_number > 1:
+                response_data['pagination']['previous'] = f'{BASE_PATH}/?offset={offset}&limit={limit}&page_number={page_number-1}&page_size={page_size}'
+                
+            if end < total_data:
+                response_data['pagination']['next'] = f'{BASE_PATH}/?offset={offset}&limit={limit}&page_number={page_number+1}&page_size={page_size}'
+                 
+        elif type(data) == dict:
+            response_data = {data} if 'error' not in data else data
+            
+        else:
+            response_data = data
+            
+        return response_data 
+             
     def get_object_list(self, session:Session = Depends(get_session), offset: int = 0, limit: int = 100):
-       
+           
         object_ = session.query(model).offset(offset).limit(limit).all()
         return object_
     
+
     def get_object(self, id:int, session:Session = Depends(get_session)):
         
         object_ = session.query(model).get(id)
@@ -19,22 +58,20 @@ class ProductsUseCases(repository):
         if object_:
             return object_ 
         else:
-            message = "Resource not found"
+            message = f"Resource with id {id} not found"
             log_api.warning(message)
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=message
-            )
             
+            response = {
+                        'error': {
+                            'code': status.HTTP_404_NOT_FOUND,
+                            'message': message
+                        }
+            }
+            
+            return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content=response)
+                        
     def add_object(self, entity:post_schema, session:Session = Depends(get_session)):
-        
-        """_summary_
-        entity.dict(exclude_unset=True) -> convierte la entidad en un diccionario
-        model(**object_data) -> crea la instancia del objeto usando el diccionario con los 2 asteriscos previos, **object_data
-        
-        Returns:
-            _type_: _description_
-        """
+
         object_data = entity.dict(exclude_unset=True)
         object_ = model(**object_data)
 
@@ -44,14 +81,7 @@ class ProductsUseCases(repository):
         return object_
 
     def add_object_list(self, entity:List[post_schema], session:Session = Depends(get_session)):
-        
-        """_summary_
-        entity.dict(exclude_unset=True) -> convierte la entidad en un diccionario
-        model(**object_data) -> crea la instancia del objeto usando el diccionario con los 2 asteriscos previos, **object_data
-        
-        Returns:
-            _type_: _description_
-        """
+
         list_objects_ = []
         
         for entity_object in entity:
@@ -65,31 +95,18 @@ class ProductsUseCases(repository):
         return list_objects_
     
     def __update_rows__(self, id:int, entity:object, session:Session = Depends(get_session)):
-        """_summary_
-        esta implementacion funciona tanto para UPDATE y PATCH dado que con el for declarado
-        se garantiza que se actualicen solo los vlores enviados
-        Args:
-            id (int): _description_
-            entity (complete_model): _description_
-            session (Session, optional): _description_. Defaults to Depends(get_session).
 
-        object_data = convierte el objeto ENTITY de entrada en un DICCIONARIO
-        setattr =  ACTUALIZA los campos del OBJECT_ de la consulta a la base de datos con los valores de OBJECT_DATA
-        Raises:
-            HTTPException: _description_
-
-        Returns:
-            _type_: _description_
-        """
-        
         object_ = self.get_object(id=id, session=session)
-        object_data = entity.dict(exclude_unset=True)
-            
-        for key, value in object_data.items():
-            setattr(object_, key, value)
-        session.add(object_)
-        session.commit()
-        session.refresh(object_)
+
+        if object_.__dict__.get('status_code') != 404:
+            object_data = entity.dict(exclude_unset=True)
+                
+            for key, value in object_data.items():
+                setattr(object_, key, value)
+            session.add(object_)
+            session.commit()
+            session.refresh(object_)
+
         return object_
         
     def update_object(self, id:int, entity:put_schema, session:Session = Depends(get_session)):
@@ -101,8 +118,18 @@ class ProductsUseCases(repository):
         return self.__update_rows__(id=id, entity=entity, session=session)
 
     def delete_object(self, id:int, session:Session = Depends(get_session)):
+        
         object_ = self.get_object(id=id, session=session)
+             
+        if isinstance(object_, model):
 
-        session.delete(object_)
-        session.commit()
-        session.close()
+            session.delete(object_)
+            session.commit()
+            session.close()
+            
+            # delete sqlalchemy key
+            object_.__dict__.pop('_sa_instance_state')
+            return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content=object_.__dict__)           
+            
+        else:
+            return object_
